One more thing
Made it so the enemy does not ignore colliders, though the raycast still
doesn't work for some reason.

Also there's really no point in having a max speed and acceleration for
the enemy as it will literally never decelerate from maximum speed. At
least with the current behavior. Just wanted to mention that somewhere.